Moves PerformEvent to reflex.
