+++ reflex-dom.cabal
@@ -84,8 +84,6 @@ library
     Reflex.Dom.Internal
     Reflex.Dom.Internal.Foreign
     Reflex.Dom.Location
-    Reflex.Dom.PerformEvent.Base
-    Reflex.Dom.PerformEvent.Class
     Reflex.Dom.PostBuild.Class
     Reflex.Dom.Old
     Reflex.Dom.Time
