+++ src/Reflex/Dom/Builder/Immediate.hs
@@ -19,7 +19,7 @@ module Reflex.Dom.Builder.Immediate where
 import Foreign.JavaScript.TH
 import Reflex
 import Reflex.Dom.Builder.Class
-import Reflex.Dom.PerformEvent.Class
+import Reflex.PerformEvent.Class
 import Reflex.Dom.PostBuild.Class
 import Reflex.Host.Class
