+++ src/Reflex/Dom/DynamicWriter.hs
@@ -29,7 +29,7 @@ import Data.Traversable
 import Reflex
 import Reflex.Dom.Builder.Class
 import Reflex.Dom.Class
-import Reflex.Dom.PerformEvent.Class
+import Reflex.PerformEvent.Class
 import Reflex.Dom.PostBuild.Class
 import Reflex.Host.Class
