+++ src/Reflex/Dom/Widget/Input.hs
@@ -43,7 +43,7 @@ import Reflex
 import Reflex.Dom.Builder.Class
 import Reflex.Dom.Builder.Immediate
 import Reflex.Dom.Class
-import Reflex.Dom.PerformEvent.Class
+import Reflex.PerformEvent.Class
 import Reflex.Dom.PostBuild.Class
 import Reflex.Dom.Widget.Basic
 import qualified Text.Read as T
